requirements.txt
click==8.2.1
ftfy==6.3.1
PyYAML==6.0.2
requests==2.32.4
html2text~=2025.4.15
psutil~=7.0.0